refactor: member
* move exception to member package

* add NoSuchAchievementException

* add null guards

* peel off dto at controller
