+++ backend/demo/src/main/java/com/_2cha/demo/auth/service/AuthService.java
@@ -6,11 +6,11 @@ import com._2cha.demo.auth.dto.JwtTokenPayload;
 import com._2cha.demo.auth.dto.OIDCUserProfile;
 import com._2cha.demo.auth.dto.TokenResponse;
 import com._2cha.demo.auth.strategy.oidc.OIDCStrategy;
-import com._2cha.demo.global.exception.NoSuchMemberException;
 import com._2cha.demo.global.exception.UnauthorizedException;
 import com._2cha.demo.member.domain.OIDCProvider;
 import com._2cha.demo.member.dto.MemberCredResponse;
 import com._2cha.demo.member.dto.MemberInfoResponse;
+import com._2cha.demo.member.exception.NoSuchMemberException;
 import com._2cha.demo.member.service.MemberService;
 import com._2cha.demo.util.BCryptHashingUtils;
 import com.auth0.jwt.JWT;
