--- backend/demo/src/main/java/com/_2cha/demo/global/exception/NoSuchMemberException.java
@@ -1,8 +1,0 @@
-package com._2cha.demo.global.exception;
-
-public class NoSuchMemberException extends NotFoundException {
-
-  public NoSuchMemberException() {
-    super("No Such Member", "noSuchMember");
-  }
-}
