+++ backend/demo/src/main/java/com/_2cha/demo/member/service/MemberService.java
@@ -1,6 +1,5 @@
 package com._2cha.demo.member.service;
 
-import com._2cha.demo.global.exception.NoSuchMemberException;
 import com._2cha.demo.member.domain.Achievement;
 import com._2cha.demo.member.domain.Member;
 import com._2cha.demo.member.domain.MemberAchievement;
@@ -10,15 +9,14 @@ import com._2cha.demo.member.dto.MemberCredResponse;
 import com._2cha.demo.member.dto.MemberInfoResponse;
 import com._2cha.demo.member.dto.MemberProfileResponse;
 import com._2cha.demo.member.dto.RelationshipOperationResponse;
-import com._2cha.demo.member.dto.SignUpRequest;
-import com._2cha.demo.member.dto.ToggleAchievementExposureRequest;
 import com._2cha.demo.member.dto.ToggleAchievementExposureResponse;
+import com._2cha.demo.member.exception.NoSuchAchievementException;
+import com._2cha.demo.member.exception.NoSuchMemberException;
 import com._2cha.demo.member.repository.AchievementRepository;
 import com._2cha.demo.member.repository.MemberQueryRepository;
 import com._2cha.demo.member.repository.MemberRepository;
 import com._2cha.demo.util.BCryptHashingUtils;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
@@ -48,10 +46,10 @@ public class MemberService {
    @ Commands
    ------------*/
   @Transactional
-  public MemberInfoResponse signUp(SignUpRequest dto) {
-    String hashed = BCryptHashingUtils.hash(dto.getPassword());
+  public MemberInfoResponse signUp(String email, String name, String password) {
+    String hashed = BCryptHashingUtils.hash(password);
 
-    Member member = Member.createMember(dto.getEmail(), hashed, dto.getName());
+    Member member = Member.createMember(email, hashed, name);
     memberRepository.save(member);
 
     return new MemberInfoResponse(member);
@@ -71,6 +69,7 @@ public class MemberService {
   public RelationshipOperationResponse follow(Long followerId, Long followingId) {
     Member follower = memberRepository.findById(followerId);
     Member following = memberRepository.findById(followingId);
+    if (follower == null || following == null) throw new NoSuchMemberException();
 
     follower.follow(following);
 
@@ -85,6 +84,7 @@ public class MemberService {
   public RelationshipOperationResponse unfollow(Long followerId, Long followingId) {
     Member follower = memberRepository.findById(followerId);
     Member following = memberRepository.findById(followingId);
+    if (follower == null || following == null) throw new NoSuchMemberException();
 
     follower.unfollow(following);
 
@@ -94,27 +94,28 @@ public class MemberService {
   @Transactional
   public void addAchievement(Long memberId, Long achvId) {
     Achievement achievement = achvRepository.findAchievementById(achvId);
+    if (achievement == null) throw new NoSuchAchievementException();
+
     Member member = memberRepository.findById(memberId);
+    if (member == null) throw new NoSuchMemberException();
+
     member.addAchievement(achievement);
   }
 
   @Transactional
   public List<ToggleAchievementExposureResponse> toggleAchievementsExposure(Long memberId,
-                                                                            List<ToggleAchievementExposureRequest> dtos) {
+                                                                            Map<Long, Boolean> achvExposureMap) {
     Member member = memberRepository.findById(memberId);
-    List<MemberAchievement> memAchvs = member.getAchievements();
+    if (member == null) throw new NoSuchMemberException();
 
-    Map<Long, Boolean> requestedAchvMap = new HashMap<>();
-    for (ToggleAchievementExposureRequest dto : dtos) {
-      requestedAchvMap.put(dto.getAchvId(), dto.isExposure());
-    }
+    List<MemberAchievement> memAchvs = member.getAchievements();
 
     List<ToggleAchievementExposureResponse> response = new ArrayList<>();
     memAchvs.forEach(
         memAchv -> {
           Long achvId = memAchv.getAchievement().getId();
-          if (requestedAchvMap.containsKey(achvId)) {
-            memAchv.toggleExposure(requestedAchvMap.get(achvId));
+          if (achvExposureMap.containsKey(achvId)) {
+            memAchv.toggleExposure(achvExposureMap.get(achvId));
             response.add(new ToggleAchievementExposureResponse(achvId, memAchv.isExposed()));
           }
         });
@@ -125,8 +126,6 @@ public class MemberService {
   /*------------
    @ Queries
    ------------*/
-
-
   public MemberProfileResponse getMemberProfileById(Long id) {
 
     MemberProfileResponse profile = memberQueryRepository.getMemberProfileById(id);
@@ -138,9 +137,7 @@ public class MemberService {
   public List<MemberProfileResponse> getMemberProfileByIdIn(List<Long> ids) {
     return memberQueryRepository.getMemberProfileByIdIn(ids);
   }
-
-  ;
-
+  
   public MemberInfoResponse getMemberInfoById(Long id) {
     MemberInfoResponse memberInfo = memberQueryRepository.getMemberInfoById(id);
     if (memberInfo == null) throw new NoSuchMemberException();
